Added 'HitAndRun' combat pattern to Grunt AI
Added the possiblity of different combat patterns for AI NPCs to follow when detecting the player.

When in the 'HitAndRun' combat pattern the AI will attempt a single attack then retreat to a random position around the player, after which they will move in for another attack and repeat the process.