Starting both portal + spirit system
The spirit system has been modified to work with the new design, however there is the need to actually implement properly the bought animations within the player character and modify its animation tree.

The portal shader has been started, it is quite ugly visually, so it might be scrapped to be re-done alltogether.