Adjustment to AI attack logic
Melee AI's now correctly raycast from their centre (previously was happening at the base of their model by mistake) when checking if the player is in front of them before deciding to attack. Helps resolve issues where the AI will stand idle when they are in attack range and the player is in front able to be hit.